add 1.setting 2.url get update 1.crawler url
