+++ parseltongue/settings.py
@@ -25,7 +25,7 @@ SECRET_KEY = '&453znkbqr(*t1m16=36ncv08k=w*$ch_c7wm033a4*qwh-l9l'
 # SECURITY WARNING: don't run with debug turned on in production!
 DEBUG = True
 
-ALLOWED_HOSTS = ['vast-refuge-25114.herokuapp.com']
+#ALLOWED_HOSTS = ['vast-refuge-25114.herokuapp.com']
 
 
 # Application definition
@@ -38,6 +38,8 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'rest_framework',
+    'crawler',
+    'setting'
 ]
 
 MIDDLEWARE = [
