--- templates/post.html
@@ -1,65 +1,0 @@
-<!DOCTYPE html>
-{% load static from staticfiles%}
-<html>
-	<head>
-		<meta charset="utf-8">
-		<title>Crawler</title>
-		<!-- Styles -->
-		<!-- Bootstrap CSS -->
-		<link type="text/css" href="{% static 'css/bootstrap.min.css' %}" rel="stylesheet">
-		<!-- Main CSS -->
-		<link type="text/css" href="{% static 'css/style-175.css' %}" rel="stylesheet">
-	</head>
-	<body>
-		<!-- UI - X Starts -->
-		<div class="ui-175">
-			<div class="container">
-				<div class="row">
-					<div>
-						<!-- Outer Starts -->
-						<div class="ui-outer">
-							<!-- Form -->
-							<div class="ui-form">
-								<!-- Heading -->
-								<h4>Carwler</h4>
-								<!-- Form -->
-								<form class="form" action="/crawler" method="post">
-									{% csrf_token %}
-									<div class="form-group">
-										<!-- html -->
-										 <textarea class="form-control" name="html" rows="10" cols="30" placeholder="Html ex:<html>...</html>"></textarea>
-									</div>
-									<div class="form-group">
-										<!-- tags -->
-										<input type="text" name="tags" class="form-control" placeholder="Text ex:SomeText or SomeText,SomeText">
-									</div>
-									<div class="form-group">
-										<!-- tags -->
-										<input type="text" name="map" class="form-control" placeholder="Text Repeat ex:{'SomeText': 2}">
-									</div>
-									<div class="form-group">
-										<!-- table -->
-										<input type="text" name="table" class="form-control" placeholder="Table Title ex:SomeTableTilte">
-									</div>
-									<div class="form-group">
-										<!-- table -->
-										<input type="text" name="json" class="form-control" placeholder="Setting json ex:{tags:{tag:[]}}">
-									</div>
-									<div class="form-group text-center">
-										<!-- submit -->
-										<button type="submit" class="btn btn-lblue">Post</button>
-									</div>
-								</form>
-							</div>
-						</div>
-						<!-- Outer Ends -->
-					</div>
-				</div>
-			</div>
-		</div>
-		<!-- UI - X Ends -->
-	<p>{{ tag }}</p>
-	<p>{{ table }}</p>
-	<p>{{ json }}</p>
-	</body>
-</html>
